+++ packages/mds-policy-service/service/helpers.ts
@@ -1,10 +1,14 @@
-import { isDefined, minutes, now, uuid } from '@mds-core/mds-utils'
+import { asArray, isDefined, minutes, now, uuid } from '@mds-core/mds-utils'
 import type { PolicyDomainModel } from '../@types'
-import type { IntentDraft, INTENT_TYPE } from '../@types/intents'
+import type { IntentDraft, IntentRuleUserFields, INTENT_TYPE } from '../@types/intents'
 import { BASE_POLICY_DEFAULTS, INTENT_RULE_CONSTANTS } from '../@types/intents'
 
 export const TWENTY_MINUTES = minutes(20)
 const POLICY_START_DATE_FUDGE_FACTOR = 1000
+
+function ruleFieldToRule(intent_type: INTENT_TYPE, ruleField: IntentRuleUserFields) {
+  return { rule_id: uuid(), ...INTENT_RULE_CONSTANTS[intent_type], ...ruleField }
+}
 
 export function translateIntentToPolicy<I extends INTENT_TYPE>(draft: IntentDraft<I>): PolicyDomainModel {
   /* The spec says that the start_date must exceed the published_date by 20 minutes.
@@ -22,11 +26,12 @@ export function translateIntentToPolicy<I extends INTENT_TYPE>(draft: IntentDraf
     draft.policy_fields.start_date = now() + TWENTY_MINUTES + POLICY_START_DATE_FUDGE_FACTOR
   }
 
-  const rule_constants = INTENT_RULE_CONSTANTS[draft.intent_type]
+  const rules = asArray(draft.rule_fields).map(rule_field => ruleFieldToRule(draft.intent_type, rule_field))
+
   return {
     policy_id: uuid(),
     ...BASE_POLICY_DEFAULTS,
     ...draft.policy_fields,
-    rules: [{ rule_id: uuid(), ...rule_constants, ...draft.rule_fields }]
+    rules
   }
 }