+++ container-images/mds-compliance-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-compliance-service
 
+## 0.2.25
+
+### Patch Changes
+
+- @mds-core/mds-compliance-service@0.10.56
+
 ## 0.2.24
 
 ### Patch Changes