+++ container-images/mds-policy-service/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @container-images/mds-policy-service
 
+## 0.2.23
+
+### Patch Changes
+
+- Updated dependencies [79df478d]
+- Updated dependencies [68fd0f77]
+- Updated dependencies [bf8c7aa5]
+  - @mds-core/mds-policy-service@0.14.0
+
 ## 0.2.22
 
 ### Patch Changes