+++ container-images/mds-policy/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @container-images/mds-policy
 
+## 0.3.24
+
+### Patch Changes
+
+- @mds-core/mds-policy@0.8.23
+
 ## 0.3.23
 
 ### Patch Changes